+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -7,7 +7,6 @@ import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.table.creator.OrderTableCreator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
